requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
Werkzeug==0.16.0
sqlalchemy>=2.0.0
python-dotenv==1.0.0
boto3>=1.28.69
PyMySQL>=1.1.0